+++ src/PageSelector.module.css
@@ -1,0 +1,2 @@
+.root {
+}
