+++ src/components/Flex.module.css
@@ -2,6 +2,10 @@
   display: flex;
 }
 
+.root[data-flex="auto"] {
+  flex: auto;
+}
+
 .root[data-direction="row"] {
   flex-direction: row;
 }
