+++ src/components/Paginator.module.css
